Add SSL support, remove binaries and fix folder structure.
