+++ LICENSE
@@ -1,6 +1,7 @@
 MIT License
 
 Copyright (c) 2016 Code Cowboy
+Copyright (c) 2018 Peter Stevenson (2E0PGS)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
