+++ SimpleSmtpClient/Form1.cs
@@ -1,0 +1,69 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Windows.Forms;
+using System.Net.Mail;
+using System.Net;
+
+namespace SimpleSmtpClient
+{
+    public partial class mainForm : Form
+    {
+        public mainForm()
+        {
+            InitializeComponent();
+        }
+
+        private void guiUseCredentials_CheckedChanged(object sender, EventArgs e)
+        {
+            guiUser.ReadOnly = true;
+            guiPassword.ReadOnly = true;
+            if (guiUseCredentials.Checked)
+            {
+                guiUser.ReadOnly = false;
+                guiPassword.ReadOnly = false;
+            }
+        }
+
+        private void guiSendMail_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                SmtpClient client = new SmtpClient();
+                client.Host = guiServerName.Text;
+                client.Port = Convert.ToInt32(guiPort.Text);
+                if (guiUseCredentials.Checked)
+                {
+                    client.UseDefaultCredentials = false;
+                    client.Credentials = new NetworkCredential(guiUser.Text, guiPassword.Text);
+                }
+                if (guiUseSsl.Checked)
+                {
+                    client.EnableSsl = true;
+                }
+                MailMessage message = CreateMailMessage();
+                client.Send(message);
+                MessageBox.Show("Email Sent.");
+
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+        }
+
+        private MailMessage CreateMailMessage()
+        {
+            MailMessage mailMessage = new MailMessage();
+            mailMessage.From = new MailAddress(guiEmailFrom.Text);
+            mailMessage.To.Add(guiEmailTo.Text);
+            mailMessage.Body = guiEmailBody.Text;
+            mailMessage.Subject = guiEmailSubject.Text;
+            return mailMessage;
+        }
+    }
+}
