--- source/SimpleSmtpClient/bin/Debug/SimpleSmtpClient.exe.config
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-<startup><supportedRuntime version="v2.0.50727"/></startup></configuration>
